% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2r.R
\name{LatexToR}
\alias{LatexToR}
\title{Translate LaTeX formula to R code (standardMath).}
\usage{
LatexToR(x, interactive = FALSE)
}
\arguments{
\item{x}{string}

\item{interactive}{boolean}
}
\value{
string
}
\description{
Take a LaTeX formula and attempts to find an equivalent representation in R.
If \code{interactive=TRUE} it pops up an interactive REPL that does the same, but interactively.
}
\examples{
LatexToR("\\\\beta_1^{\\\\frac{x+1}{x^2 \\\\cdot y}}")
}
